docs/resources/source-code/ISWC/src/Integration/GenericJob/requirements.txt
autopep8==1.5.1
azure-cosmos==3.1.2
jsonschema==3.2.0
opencensus-ext-azure==1.0.2
pandas==1.3.5
paramiko==2.11.0
pysftp==0.2.9
pytest==3.8.1
setuptools==58.0.4
azure-mgmt-datafactory==0.13.0
